Disable quiesce, and thus fastreboot
Empirically, fastreboot with KVM loaded hangs before the system reboots
in circumstances which are mysterious.

It would be nice if we could detach KVM prior to reboot, but by the time
we reach quiesce(9E) we're running single-threaded on one CPU, and
detaching KVM in any meaningful way involves cross-calls to disable VMX
on all CPUs.

The safest thing to do, at least pending any real knowledge of precisely
why we're hanging up, is to refuse to quiesce.